币币账单接口 修改

+++ ex-bb-extend/src/main/java/com/hp/sh/expv3/bb/extension/service/BbAccountLogExtService.java
@@ -10,6 +10,6 @@ import java.util.List;
 public interface BbAccountLogExtService {
     List<BbAccountLogExtVo> listBbAccountLogs(Long userId, String asset,List<String> symbols, Integer historyType, Integer tradeType, Long startDate, Long endDate,  Integer pageSize);
 
-    List<BbAccountLogExtVo> listBbAccountLogsByPage(Long userId, String asset, List<String> symbols, Integer historyType, Integer tradeType, Integer lastId, Integer nextPage, Long startDate, Long endDate, Integer pageSize);
+    List<BbAccountLogExtVo> listBbAccountLogsByPage(Long userId, String asset, List<String> symbols, Integer historyType, Integer tradeType, Long lastId, Integer nextPage, Long startDate, Long endDate, Integer pageSize);
 
 }
